+++ docs/web/@types/global.d.ts
@@ -3,9 +3,6 @@ import { IWindow } from 'ssr-types'
 declare module '*.less'
 
 declare global {
-  interface Window {
-    __USE_SSR__?: IWindow['__USE_SSR__']
-    __INITIAL_DATA__?: IWindow['__INITIAL_DATA__']
-  }
+  interface Window extends IWindow {}
   const __isBrowser__: Boolean
 }