+++ example/nestjs-vue3-ssr/web/@types/global.d.ts
@@ -1,11 +1,7 @@
 import { IWindow } from 'ssr-types'
 
 declare global {
-  interface Window {
-    __USE_SSR__?: IWindow['__USE_SSR__']
-    __USE_VITE__?: boolean
-    __INITIAL_DATA__?: IWindow['__INITIAL_DATA__']
-  }
+  interface Window extends IWindow {}
   const __isBrowser__: Boolean
 }