+++ packages/client-utils/src/index.ts
@@ -1,7 +1,6 @@
 /* utils */
+export * from 'ssr-common-utils'
 export * from './preload'
-export * from './findRoute'
-export * from './utils'
 
 /* types */
 export * from 'ssr-types'