+++ packages/client-utils/src/preload.ts
@@ -1,7 +1,6 @@
 import { ReactESMPreloadFeRouteItem, ReactESMFeRouteItem } from 'ssr-types-react'
+import { normalizePath } from 'ssr-common-utils'
 import { pathToRegexp } from 'path-to-regexp'
-import { normalizePath } from './utils'
-
 const preloadComponent = async (Routes: ReactESMPreloadFeRouteItem[], base?: string) => {
   for (const route of Routes) {
     const { component, path } = route