+++ packages/hoc-react/src/wrapComponent.tsx
@@ -1,11 +1,10 @@
 import * as React from 'react'
+import 'react-router'
 import { useContext, useEffect, useState } from 'react'
 import { withRouter, RouteComponentProps } from 'react-router-dom'
-import { DynamicFC, StaticFC, Action, IWindow, ReactESMFetch, ReactFetch } from 'ssr-types-react'
+import { DynamicFC, StaticFC, Action, ReactESMFetch, ReactFetch } from 'ssr-types-react'
 // @ts-expect-error
 import { STORE_CONTEXT } from '_build/create-context'
-
-declare const window: IWindow
 
 let hasRender = false