+++ packages/plugin-react/src/entry/client-entry.tsx
@@ -1,22 +1,20 @@
 import * as React from 'react'
 import * as ReactDOM from 'react-dom'
 import { BrowserRouter, Route, Switch } from 'react-router-dom'
-import { preloadComponent } from 'ssr-client-utils'
+import { preloadComponent, isMicro } from 'ssr-client-utils'
 import { wrapComponent } from 'ssr-hoc-react'
-import { IWindow, LayoutProps, ReactRoutesType } from 'ssr-types-react'
+import { LayoutProps, ReactRoutesType } from 'ssr-types-react'
 import { Routes } from './create-router'
 import { AppContext } from './context'
 
 const { FeRoutes, layoutFetch, App } = Routes as ReactRoutesType
-
-declare const window: IWindow
 
 const clientRender = async (): Promise<void> => {
   const IApp = App ?? function (props: LayoutProps) {
     return props.children!
   }
   // 客户端渲染||hydrate
-  const baseName = window.microApp ? window.clientPrefix : window.prefix
+  const baseName = isMicro() ? window.clientPrefix : window.prefix
   const routes = await preloadComponent(FeRoutes, baseName)
   ReactDOM[window.__USE_SSR__ ? 'hydrate' : 'render'](
     <BrowserRouter basename={baseName}>