+++ packages/types/src/config.ts
@@ -1,5 +1,6 @@
 import { Options, RuleSetCondition } from 'webpack'
 import * as Config from 'webpack-chain'
+import type { PluginOption, ServerOptions, UserConfig as ViteConfig } from 'vite'
 import type { RollupBabelInputPluginOptions } from '@rollup/plugin-babel'
 import { Argv } from './yargs'
 import { ISSRContext } from './ctx'
@@ -47,7 +48,7 @@ export interface IConfig {
       }
     }
   }
-  chainBaseConfig: (config: Config) => void
+  chainBaseConfig: (config: Config, isServer: boolean) => void
   chainServerConfig: (config: Config) => void
   chainClientConfig: (config: Config) => void
   webpackStatsOption: Options.Stats
@@ -90,17 +91,24 @@ export interface IConfig {
   supportOptinalChaining: boolean
   viteConfig?: () => {
     common?: {
-    // 双端通用插件
-      extraPlugin?: any[]
+      // 双端通用配置
+      extraPlugin?: PluginOption | PluginOption[]
+      server?: ServerOptions
     }
     client?: {
-      defaultPluginOptions?: any
-      extraPlugin?: any[]
+      defaultPluginOptions?: any // 为默认装载的插件定义 options, vue3 场景是 @vitejs/plugin-vue, react 场景是 @vitejs/plugin-react
+      extraPlugin?: PluginOption | PluginOption[]
+      otherConfig?: ViteConfig
     }
     server?: {
       defaultPluginOptions?: any
-      extraPlugin?: any[]
+      extraPlugin?: PluginOption | PluginOption[]
+      otherConfig?: ViteConfig
     }
+  }
+  hmr?: {
+    host?: string
+    port?: number
   }
   define?: {
     base?: Record<string, string>